Remove SidechainMessageRecoveryCommand & verifyMessageRecovery
